fix: correct (short) cast to (int) in BufferUtils.putUnsignedInt (Issue 26)
putUnsignedInt(ByteBuffer, int, long) was casting the masked long value
to (short) before passing it to ByteBuffer.putInt(), silently truncating
any value above 0x7FFF to 16 bits. Changed the cast to (int) so the
full 32-bit unsigned value is preserved.

+++ AndEngine-GLES2/src/org/andengine/opengl/util/BufferUtils.java
@@ -117,7 +117,7 @@ public class BufferUtils {
 	}
 
 	public static void putUnsignedInt(final ByteBuffer pByteBuffer, final int pPosition, final long pValue) {
-		pByteBuffer.putInt(pPosition, (short) (pValue & 0xFFFFFFFFL));
+		pByteBuffer.putInt(pPosition, (int) (pValue & 0xFFFFFFFFL));
 	}
 
 	// ===========================================================
